requirements.txt
Flask
flask-cors
flask-pymongo
python-dotenv
bcrypt
PyJWT
psutil
langchain_groq
langchain_core
langchain_community
langchain_huggingface
pymongo
# error_handler
faiss-cpu
gunicorn 
nltk
Authlib
spacy
scikit-learn
schedule
pytz
textblob
cryptography
google-auth
afinn
SMTPEmail
email-to
APScheduler
pandas
openpyxl
gspread 
oauth2client
sentence-transformers
twilio